pyproject.toml
[tool.poetry]
name = "rkeeper-adapter"
version = "0.1.0"
description = ""
authors = ["adonskoi <donskoy.alexander@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
httpx = "^0.22.0"
celery = "^5.2.6"
redis = "^4.2.2"
types-redis = "^4.1.21"
python-dotenv = "^0.20.0"
docker = "^5.0.3"
click = "^8.1.2"
uvicorn = "^0.17.6"
structlog = "^21.5.0"
pre-commit = "^2.20.0"
opentelemetry-distro = "^0.36b0"
opentelemetry-exporter-jaeger = "^1.15.0"
opentelemetry-instrumentation-fastapi = "^0.36b0"
opentelemetry-instrumentation-httpx = "^0.36b0"
opentelemetry-instrumentation-celery = "^0.36b0"
starter-dto = "^0.2.8"
opentelemetry-propagator-jaeger = "^1.15.0"
types-pytz = "^2023.3.1.1"
gunicorn = "^21.2.0"
fastapi = "0.99.1"
alembic = "^1.13.1"
psycopg2-binary = "^2.9.9"
pydantic = "1.10.0a1"
sqlalchemy = "2.0.23"
ipython = "^8.23.0"
mypy = "1.10.0"
pytest-cov = "^5.0.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
black = "^22.3.0"

[[tool.poetry.source]]
name = "gitlab"
url = "https://gitlab.handh.ru/api/v4/projects/127/packages/pypi/simple"


[tool.poetry.group.dev.dependencies]
bandit = "^1.8.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
ignore_missing_imports = true
disallow_untyped_defs = true
exclude = "tests/"

[tool.black]
line_length = 120
